+++ Resources/config/services.yml
@@ -4,12 +4,13 @@ parameters:
     zikula_dizkus_module.twig.extension.class: Zikula\DizkusModule\Twig\TwigExtension
     zikula_dizkus_module.links.class: Zikula\DizkusModule\Container\LinkContainer
 
-    zikula_dizkus_module.forum_user_helper.class: Zikula\DizkusModule\Helper\ForumUserHelper
     zikula_dizkus_module.moderators_helper.class: Zikula\DizkusModule\Helper\ModeratorsHelper
     zikula_dizkus_module.favorites_helper.class: Zikula\DizkusModule\Helper\FavoritesHelper
     zikula_dizkus_module.count_helper.class: Zikula\DizkusModule\Helper\CountHelper
     zikula_dizkus_module.synchronization_helper.class: Zikula\DizkusModule\Helper\SynchronizationHelper
+    zikula_dizkus_module.rank_helper.class: Zikula\DizkusModule\Helper\RankHelper
 
+    zikula_dizkus_module.forum_user_manager.class: Zikula\DizkusModule\Manager\ForumUserManager
     zikula_dizkus_module.forum_manager.class: Zikula\DizkusModule\Manager\ForumManager
     zikula_dizkus_module.topic_manager.class: Zikula\DizkusModule\Manager\TopicManager
     zikula_dizkus_module.post_manager.class: Zikula\DizkusModule\Manager\PostManager
@@ -22,7 +23,12 @@ services:
         class: %zikula_dizkus_module.security.class%
         tags:
             - { name: zikula.dizkus.security }
-        arguments: [@request_stack, @doctrine.orm.default_entity_manager, @translator.default, @zikula_permissions_module.api.permission, @zikula_extensions_module.api.variable]
+        arguments: 
+            - "@request_stack"
+            - "@doctrine.orm.default_entity_manager"
+            - "@translator.default"
+            - "@zikula_permissions_module.api.permission"
+            - "@zikula_extensions_module.api.variable"
         
     zikula_dizkus_module.twig.extension:
         class: %zikula_dizkus_module.twig.extension.class%
@@ -44,13 +50,7 @@ services:
         arguments: [@request_stack, @doctrine.orm.default_entity_manager, @router]
 
 # Helpers
-        
-    zikula_dizkus_module.forum_user_helper:
-        class: %zikula_dizkus_module.forum_user_helper.class%
-        tags:
-            - { name: zikula.dizkus.forum_user_helper }
-        arguments: [@request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @translator.default]
-  
+          
     zikula_dizkus_module.favorites_helper:
         class: %zikula_dizkus_module.favorites_helper.class%
         tags:
@@ -74,23 +74,71 @@ services:
         tags:
             - { name: zikula.dizkus.synchronization_helper }
         arguments: [@doctrine.orm.default_entity_manager, @zikula_dizkus_module.count_helper]
- 
-# Managers
+         
+    zikula_dizkus_module.rank_helper:
+        class: %zikula_dizkus_module.rank_helper.class%
+        tags:
+            - { name: zikula.dizkus.rank_helper }
+        arguments: [@doctrine.orm.default_entity_manager, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable, @zikula_dizkus_module.forum_user_manager]
         
+# Managers
+
+    zikula_dizkus_module.forum_user_manager:
+        class: %zikula_dizkus_module.forum_user_manager.class%
+        tags:
+            - { name: zikula.dizkus.forum_user_manager }
+        arguments:
+            - "@translator.default"
+            - "@router"
+            - "@request_stack"
+            - "@doctrine.orm.default_entity_manager"
+            - "@zikula_users_module.current_user"
+            - "@zikula_dizkus_module.security"
+            - "@zikula_extensions_module.api.variable"
+            - "@zikula_permissions_module.api.permission"
+
     zikula_dizkus_module.forum_manager:
         class: %zikula_dizkus_module.forum_manager.class%
         tags:
             - { name: zikula.dizkus.forum_manager }
-        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable, @zikula_permissions_module.api.permission]
+        arguments:
+            - "@translator.default"
+            - "@router"
+            - "@request_stack"
+            - "@doctrine.orm.default_entity_manager"
+            - "@zikula_users_module.current_user"
+            - "@zikula_dizkus_module.security"
+            - "@zikula_extensions_module.api.variable"
+            - "@zikula_permissions_module.api.permission"
 
     zikula_dizkus_module.topic_manager:
         class: %zikula_dizkus_module.topic_manager.class%
         tags:
             - { name: zikula.dizkus.topic_manager }
-        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable, @zikula_dizkus_module.forum_manager, @zikula_dizkus_module.synchronization_helper]
-
+        arguments: 
+            - "@translator.default"
+            - "@router"
+            - "@request_stack"
+            - "@doctrine.orm.default_entity_manager"
+            - "@zikula_users_module.current_user"
+            - "@zikula_dizkus_module.security"
+            - "@zikula_extensions_module.api.variable"
+            - "@zikula_dizkus_module.forum_user_manager"
+            - "@zikula_dizkus_module.forum_manager"
+            - "@zikula_dizkus_module.synchronization_helper"        
+        
     zikula_dizkus_module.post_manager:
         class: %zikula_dizkus_module.post_manager.class%
         tags:
             - { name: zikula.dizkus.post_manager }
-        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable]
+        arguments: 
+            - "@translator.default"
+            - "@router"
+            - "@request_stack"
+            - "@doctrine.orm.default_entity_manager"
+            - "@zikula_users_module.current_user"
+            - "@zikula_dizkus_module.security"
+            - "@zikula_extensions_module.api.variable"
+            - "@zikula_dizkus_module.forum_manager"
+            - "@zikula_dizkus_module.topic_manager"
+            - "@zikula_dizkus_module.synchronization_helper"
